Gjesteland Poisson Debug/Sort
Added a method debug_poisson in tgfn1_peakfind to find a Gjesteland
event that was lost before. Will generalize it to be its own procedure
next week.

For the binary microsecond sensitivity we cannot use uniq unless we
also sort the times. They are not always sorted (see Infanger Slides
08/01/2015).

Only config changes are comments.